added root types folder to contain general interfaces
